% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_wrapper.R
\name{dectree_wrapper}
\alias{dectree_wrapper}
\title{The wrapper function of the decision tree model exported from AMUA}
\usage{
dectree_wrapper(params, treefunc, popsize = 1)
}
\arguments{
\item{params}{A list of parameters with the basecase values. The list should be organized
as `list(par1 = 0.5, par2 = 0.2)` with the parameter names and values. For the AMUA
decision tree model, this argument is the 1st output of `txt_parse()`.}

\item{treefunc}{Text data of the AMUA decision model. This argument is the 2nd output of `txt_parse()`.}

\item{popsize}{The population size of the decision problem. The default is 1.}
}
\value{
The function returns a `data.frame` with the strategy name as the first column and the
        expected effectiveness and cost are in the following columns.
}
\description{
This is the wrapper function of the decision tree movel exported from AMUA.
             It is required to run `txt_parse()` first before running this function.
}
