% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_gen_psa.R
\name{gen_psa_samp_code}
\alias{gen_psa_samp_code}
\title{Generate code of `gen_psa_samp()` in `dampack`}
\usage{
gen_psa_samp_code(param_table, n_samp = 100)
}
\arguments{
\item{param_table}{a `data.frame` including information of the parameter names, the distribution,
and the required parameters for the distributions.}

\item{n_samp}{number of samples for `gen_psa_samp()`}
}
\value{
The function returns the `R` code for `gen_psa_samp()`.
}
\description{
This function will generate R code to run `gen_psa_samp()` in `dampack`
}
