% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_wrapper.R
\name{markov_decision_wrapper}
\alias{markov_decision_wrapper}
\title{Wrapper function running a Markov model for each strategy.}
\usage{
markov_decision_wrapper(
  vary_param_ls = NULL,
  other_input_ls = NULL,
  userfun,
  strategy_set
)
}
\arguments{
\item{vary_param_ls}{This argument takes a list of parameters that are varied in deterministic
or probabilistic sensitivity analyses. The list should be named and the length of each element
should be 1. For example, `list(par1 = 0.2, par2 = 0.3, par3 = 0.2)`.}

\item{other_input_ls}{The argument is a named list. This argument takes parameters or data that do not
change in deterministic or probablistic sensitivity analysis.}

\item{userfun}{This is a user defined markov model function. See the example function `markov_model()`.}

\item{strategy_set}{The set of strategies that are embedded in the `markov_model()`.}
}
\value{
The function returns a `data.frame` with first column as the strategies, the other columns are
        the outcomes of interest.
}
\description{
This is a wrapper function to calculate the outcome for every strategy.
}
