% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_wrapper.R
\name{markov_model}
\alias{markov_model}
\title{Example of Markov model function}
\usage{
markov_model(l_param_all, strategy = NULL)
}
\arguments{
\item{l_param_all}{All parameter inputs. This is a named list organized as
`list(par1 = 0.3, par2 = 0.5, df1 = data.frame(t = c(1:3), param = c(0.2, 0.2, 0.3)))`}

\item{strategy}{The strategy of interest. The default is `NULL`, which is the do nothing strategy.}
}
\value{
The function returns a `data.frame` with first column as the strategy of interest,
        and the other columns are the outcomes of interest.
}
\description{
This Markov model function using Ontario tanning ban as an example
}
