% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_wrapper.R
\name{parse_amua_tree}
\alias{parse_amua_tree}
\title{Parsing text file of the decision tree model exported from AMUA}
\usage{
parse_amua_tree(path_to_txt)
}
\arguments{
\item{path_to_txt}{A string of the path to the decision tree model exported from AMUA.}
}
\value{
The function returns a list of three elements: `param_names`, `param_val`, and `treefunc`.
    \item{param_ls}{A list of parameters with parameter names.}
    \item{treefunc}{Text of the decision tree exported from AMUA.}
}
\description{
This function is to parse the decision tree model exported from AMUA. The function returns
             a list of three elements: `param_names`, `param_val`, and `treefunc`.
}
